+++ packages/shared/src/types/index.ts
@@ -1,3 +1,3 @@
 export * from './spec';
 export * from './condition';
-export * from './utils';
+export * from './schema';
